Feat: accessLink를 LinkCommand로 변경 후 비동기 적용 및 테스트 수정
